refactor header and add new feature for saving changes
