+++ src/App.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import './App.css';
-import { Header } from './components/header';
+import { Header } from './components/header/header';
 import { AccessManagerContainer } from './components/access-manager-container';
 
 function App() {
